// types/cfn-resources/aws-lambda-layerversionpermission.d.ts
/**
 * TypeScript definitions for AWS::Lambda::LayerVersionPermission
 * Generated from CloudFormation Resource Schema
 */

/**
 * Schema for Lambda LayerVersionPermission
 */
export namespace AWS_Lambda_LayerVersionPermission {
  /**
   * Schema for Lambda LayerVersionPermission
   */
  export type LayerVersionPermissionResourceType = {
    /**
     * ID generated by service
     * Read-only property
     */
    Id?: string;

    /**
     * The API action that grants access to the layer.
     * Create-only property
     */
    Action: string;

    /**
     * The name or Amazon Resource Name (ARN) of the layer.
     * Create-only property
     */
    LayerVersionArn: string;

    /**
     * With the principal set to *, grant permission to all accounts in the specified organization.
     * Create-only property
     */
    OrganizationId?: string;

    /**
     * An account ID, or * to grant layer usage permission to all accounts in an organization, or all AWS accounts (if organizationId is not specified).
     * Create-only property
     */
    Principal: string;
  };
}